fix: cache pane-border git info to stop split-window tmux wedge
The pane-border-format ran #() git shell-jobs per pane; a split-window on
a window with two size-mismatched clients re-fired them in a resize storm
that wedged the whole tmux server ~40-64s, freezing the run-kit UI.

Extract the git-branch resolver into internal/gitinfo (shared with the web
sidebar), have the daemon snapshotter stamp @rk_pane_git_branch/
@rk_pane_git_worktree/@rk_pane_pathtail per pane only-on-change, and rewrite
pane-border-format to read those options. No #() jobs remain, so the fork
storm cannot occur.
