+++ configs/tmux/default.conf
@@ -72,7 +72,7 @@ set -g pane-active-border-style "fg=colour4"
 
 # Pane status bar — shows command, path, git branch, worktree badge
 set -g pane-border-status top
-set -g pane-border-format '#{?pane_active,#[bg=colour4#,fg=colour15#,bold]  #{?#{e|>:#{window_panes},1},#P · ,}#(echo #{pane_current_path} | rev | cut -d/ -f1-2 | rev) #[bg=colour3#,fg=colour15#,bold]#(cd #{pane_current_path} && git rev-parse --show-toplevel 2>/dev/null | grep -q "\.worktrees\|worktrees" && echo "   " || echo "")#[bg=colour8#,fg=colour7#,bold]  #(cd #{pane_current_path} && git branch --show-current 2>/dev/null || git rev-parse --short HEAD 2>/dev/null | cut -c1-20) #[bg=colour2#,fg=colour7] #{pane_current_command} ,#[fg=colour8]  #{?#{e|>:#{window_panes},1},#P · ,}#(echo #{pane_current_path} | rev | cut -d/ -f1-2 | rev) }'
+set -g pane-border-format '#{?pane_active,#[bg=colour4#,fg=colour15#,bold]  #{?#{e|>:#{window_panes},1},#P · ,}#{@rk_pane_pathtail} #[bg=colour3#,fg=colour15#,bold]#{@rk_pane_git_worktree}#[bg=colour8#,fg=colour7#,bold]  #{@rk_pane_git_branch} #[bg=colour2#,fg=colour7] #{pane_current_command} ,#[fg=colour8]  #{?#{e|>:#{window_panes},1},#P · ,}#{@rk_pane_pathtail} }'
 
 # --- Keybindings (byobu-style) ---
 bind -n F2 new-window
