+++ docs/memory/run-kit/configuration.md
@@ -32,7 +32,9 @@ When the resolved config path is not `DefaultConfigPath` (the `tmux_conf` key or
 
 Every scaffold path — `EnsureConfig`, `ForceWriteConfig`, `rk mux init-conf` (both cobra instances), `POST /api/tmux/init-conf` — also ensures `tmux.d/` exists and scaffolds `tmux.d/user.conf` as a commented starter (purpose, one commented `set -g` example, the `10-*.conf` numeric-ordering pointer) when absent; an existing `user.conf` is never overwritten, including under `--force`. (0tu6)
 
-The embedded conf's `pane-border-format` renders, in both the active and inactive arms, a conditional pane-index prefix `#{?#{e|>:#{window_panes},1},#P · ,}` immediately before the path segment (the last two segments of `pane_current_path`), followed by the worktree badge, git branch, and `pane_current_command` segments with per-state color arms: a single-pane window shows the bare path, a multi-pane window shows `N · <path>`. The raw `%N` pane id appears nowhere in the border — it stays reachable via the web status bar's `Copy tmux pane id` affordance and `tmux display-message -p '#{pane_id}'`. (qt7k)
+The embedded conf's `pane-border-format` renders, in both the active and inactive arms, a conditional pane-index prefix `#{?#{e|>:#{window_panes},1},#P · ,}` immediately before the path segment, followed by the worktree badge, git branch, and `pane_current_command` segments with per-state color arms: a single-pane window shows the bare path, a multi-pane window shows `N · <path>`. The raw `%N` pane id appears nowhere in the border — it stays reachable via the web status bar's `Copy tmux pane id` affordance and `tmux display-message -p '#{pane_id}'`. (qt7k)
+
+The path segment, worktree badge, and git branch are **read from daemon-stamped pane user-options** — `#{@rk_pane_pathtail}`, `#{@rk_pane_git_worktree}`, `#{@rk_pane_git_branch}` — so the format contains **no `#()` shell jobs**: drawing a pane border forks no subprocess. The daemon (snapshotter) stamps those options; the values match the web sidebar's git branch because both derive from the same `internal/gitinfo` resolver. See [agent-state](/run-kit/agent-state.md) § the `@rk_pane_git_*` options for the stamping contract. An unstamped pane (brand-new, or a server the daemon just began covering) renders those segments empty until the next stamp — no fork, no fallback. (s9fw)
 
 ## Settings Registry
 
@@ -239,5 +241,11 @@ Every scaffold path (`EnsureConfig`, `ForceWriteConfig`, `rk mux init-conf`, `PO
 **Why**: the single-pane window is the dominant case — a constant `1 · ` prefix is noise that duplicates nothing useful, and a raw tmux internal id means nothing to a user reading primary chrome. The id stays reachable where it belongs: the web status bar's `Copy tmux pane id` affordance and `tmux display-message -p '#{pane_id}'`.
 **Rejected**: keeping `#P` unconditionally (noise in the common case); keeping `#D` anywhere in the border (the internal-id leak).
 *Introduced by*: 260902-qt7k-ipad-chrome-polish
+
+### Pane-border git info is daemon-stamped, not forked at draw time
+**Decision**: the `pane-border-format` reads its path tail, worktree badge, and git branch from pane user-options (`@rk_pane_pathtail` / `@rk_pane_git_worktree` / `@rk_pane_git_branch`) that the daemon's snapshotter stamps only-on-change; the format runs no `#()` shell jobs.
+**Why**: the old format ran `#()` git shell jobs (`git rev-parse`, `git branch`) per pane on every border draw. On a `split-window` where a window has two size-participating clients of different sizes (the common case: an agent's own terminal plus the run-kit web viewer), the resize renegotiation re-fires those jobs in a burst that wedges the whole tmux server ~40–64s, freezing every client and the dashboard. Caching the values as options the daemon computes (via the shared `internal/gitinfo` resolver, so the border matches the sidebar) removes every draw-time fork, so the storm cannot occur regardless of how many resizes happen. Stamping lives on the snapshotter tick because it is subscriber-independent — a native-terminal-only user with no browser open still gets a populated border.
+**Rejected**: stamping from the SSE poll (browser-gated — blanks the border when no web client watches the server); a `window-size`/`aggressive-resize` change to suppress the renegotiation (alters the multi-viewer sizing behavior run-kit is built around); dropping the branch/badge from the border entirely (loses information users rely on).
+*Introduced by*: 260908-s9fw-cache-pane-border-git-info
 
 See [architecture](/run-kit/architecture.md) § `internal/settings` for the package-level contract and [layout-snapshots](/run-kit/layout-snapshots.md) for the snapshot store under the state root.
