+++ fab/changes/260908-s9fw-cache-pane-border-git-info/.history.jsonl
@@ -1,0 +1,12 @@
+{"action":"enter","driver":"fab-new","event":"stage-transition","stage":"intake","ts":"2026-09-08T19:25:12Z"}
+{"args":"Fix tmux pane-border-format wedge: replace per-redraw #() git shell jobs with a cached pane user-option stamped by the backend (which already derives git branch), so split-window with two size-mismatched clients no longer stalls the tmux server ~40-64s and freezes the run-kit UI","cmd":"fab-new","event":"command","ts":"2026-09-08T19:25:12Z"}
+{"delta":"+4.8","event":"confidence","score":4.8,"trigger":"calc-score","ts":"2026-09-08T19:27:16Z"}
+{"cmd":"fab-ff","event":"command","ts":"2026-09-08T19:28:00Z"}
+{"action":"enter","driver":"fab-ff","event":"stage-transition","stage":"apply","ts":"2026-09-08T19:28:00Z"}
+{"action":"enter","driver":"fab-ff","event":"stage-transition","stage":"review","ts":"2026-09-08T19:59:49Z"}
+{"event":"review","result":"failed","ts":"2026-09-08T20:06:45Z"}
+{"action":"re-entry","driver":"fab-ff","event":"stage-transition","stage":"apply","ts":"2026-09-08T20:06:45Z"}
+{"action":"re-entry","driver":"fab-ff","event":"stage-transition","stage":"review","ts":"2026-09-08T20:07:59Z"}
+{"action":"enter","driver":"fab-ff","event":"stage-transition","stage":"hydrate","ts":"2026-09-08T20:10:34Z"}
+{"event":"review","result":"passed","ts":"2026-09-08T20:10:34Z"}
+{"action":"enter","driver":"fab-ff","event":"stage-transition","stage":"ship","ts":"2026-09-08T20:12:40Z"}
